tests for item were added
